pyproject.toml
[project]
name = "web-posts"
version = "0.1.0"
description = ""
authors = [
    {name = "dim4ik85",email = "dim4ik020985@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.12,<4"
dependencies = [
    "django (>=5.1.7,<6.0.0)",
    "djangorestframework (>=3.15.2,<4.0.0)",
    "psycopg2-binary (>=2.9.10,<3.0.0)",
    "drf-yasg (>=1.21.10,<2.0.0)",
    "django-environ (>=0.12.0,<0.13.0)",
    "pillow (>=11.1.0,<12.0.0)"
]

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.group.dev.dependencies]
pre-commit = "^4.2.0"
black = "^25.1.0"
flake8 = "^7.1.2"
flake8-bugbear = "^24.12.12"
isort = "^6.0.1"
mypy = "^1.15.0"
